+++ pyogrio/_compat.py
@@ -33,6 +33,8 @@ PANDAS_GE_15 = pandas is not None and Version(pandas.__version__) >= Version("1.
 PANDAS_GE_20 = pandas is not None and Version(pandas.__version__) >= Version("2.0.0")
 PANDAS_GE_22 = pandas is not None and Version(pandas.__version__) >= Version("2.2.0")
 
+GDAL_GE_38 = __gdal_version__ >= (3, 8, 0)
+
 HAS_GDAL_GEOS = __gdal_geos_version__ is not None
 
 HAS_SHAPELY = shapely is not None and Version(shapely.__version__) >= Version("2.0.0")